requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.5.1
tensorflow-cpu==2.12.0
xgboost==2.0.3
joblib==1.4.2
openpyxl==3.1.5